update app module to app config and update providers
